feat(workers|coverage): add #[repr(u8)] for structs & add tests

+++ src/workers/types/listener.rs
@@ -15,7 +15,7 @@ pub async fn run_dns_listener(
 ) -> Result<(), SCloudException> {
     let socket = UdpSocket::bind(bind_addr).await.map_err(|_| SCloudException::SCLOUD_WORKER_LISTENER_BIND_FAILED)?;
     let mut buf = [0u8; 65_535];
-    worker.set_state(WorkerState::IDLE as u8);
+    worker.set_state(WorkerState::IDLE);
 
     loop {
         let (len, src) = socket.recv_from(&mut buf).await.map_err(|_| SCloudException::SCLOUD_WORKER_LISTENER_RECV_FAILED)?;
